javasparrow-server/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-allauth==0.33.0
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-auth==0.9.2
django-toolbelt==0.0.1
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
flake8==3.4.1
gunicorn==19.7.1
idna==2.6
Markdown==2.6.9
mccabe==0.6.1
oauthlib==2.0.4
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
pycodestyle==2.3.1
pyflakes==1.5.0
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
